Types Packages
Starting to add type validation for the JS version of things. Once things are more strongly-typed, then I can move the rest over to TS!

Looking to possible make a package of my own that merges a fork of all of these into a single API, which can be represented by a single intermediate format interface structure, that may be a bit too complex though, we'll see. I also like the idea of just converting between each format in it's own right. That's the dilemma I'm facing for working with multi-platform Minecraft save parsing as well. Should each format work on it's own, then you convert between them separately, based on your specific needs, or should it all go through a central format that represents everything all in one go? That t'is the question.

"Transcentience", listening to Animals as Leaders today.