npm audit fix --force
Still seems a bit broken, trying to resolve the package issues. Running both `npm run build` and `npm run test` didn't work, and I want to ensure the current state of the code passes the test, before I start modifying it. It's just that the dependencies tend to fall out of whack after a long time, seems like this happens with most projects.