ESM Debugging
This project isn't quite ready for ESM yet, so when upgrading the dev dependencies to their latest versions, they got the ESM builds, which didn't work with Gulp here. I went back a version or two with they are still in CJS, just to enable the tests to work for the meantime. I want to get the tooling working before trying to move the code over as well.

I've noticed with other projects I've tried migrating before, that if you don't migrate/fix the tooling first as well, then it's going to bite you later. Make sure the existing tests still work, for your own sake honestly. Even if you don't use tests normally, or you just want to use ESM, you still want to migrate things properly. My concept here is that everything should migrate losslessly, in terms of ensuring to preserve the same logic of the original state of the library before you started working with it.

Rush - Limbo

https://github.com/sindresorhus/gulp-mocha/releases/tag/v10.0.0
https://github.com/sindresorhus/del/releases/tag/v7.0.0