TSC Migration
Moved the build step over to TSC, away from Gulp and Babel. Other than JSHint removal, everything should essentially be the same as before. I also made it so the dist folder doesn't get committed to the repo, usually that's avised against doing that, so I made that change as well.

I still have to rehash the `example*` commands, I was going to do that in another step.