TS Renaming
Was going to try typing things out with JSDoc like I have done initially for other projects, but I think this one will require a bit more complex of types to do that, so I'm going to use full TS for it's entire feature set instead. I didn't want to deal with generics for functions in JSDoc.