Improved TS Module Resolution
This is to just help with TS things in general, as well as for when I move things to full ESM down the road.

I need to migrate the user of Node's Buffer over to plain TypedArrays and DataViews as well, then it won't need a polyfill for use in the browser.