Further Internal Types
Working on typing the library from the inside-out. Glad I went with full TS, the generics helped out a bunch!

Got some of the value types' information from the README, accidentally discovered it when searching for the various symbols across the source. Forgot to check the README for any documentation as well.

https://www.youtube.com/watch?v=meEfdkQtQwg
https://www.youtube.com/watch?v=PlbjcTjuni8
https://www.youtube.com/watch?v=S9CLh79huBE
https://www.youtube.com/watch?v=7ROfaO9iRcs