More Internal Types
Found an already-updated version to this project it seems, so my work on this may not be fully necessary. But, we'll see. I'm not completely on board with using classes-galore to implement all of the shapes, I think there's something to be said for plain interface types, along with enum key types, where you can associate a given type value with that of it's interface shape.

https://github.com/sports-alliance/sports-lib
https://www.npmjs.com/package/@xmldom/xmldom

King's X - Dogman, Sunshine Rain